+++ packages/node/src/indexer/x-provider/cachedProvider.ts
@@ -1,0 +1,49 @@
+// Copyright 2020-2022 OnFinality Limited authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { ProviderInterface } from '@polkadot/rpc-provider/types';
+import LRUCache from 'lru-cache';
+
+const MAX_CACHE_SIZE = 200;
+const CACHE_TTL = 60 * 1000;
+
+/* eslint-disable prefer-rest-params */
+export function createCachedProvider(provider: ProviderInterface) {
+  const cacheMap = new LRUCache({ max: MAX_CACHE_SIZE, ttl: CACHE_TTL });
+
+  const cachedMethodHandler = (method, params, target, args) => {
+    const cacheKey = `${method}-${params[0]}`;
+    if (cacheMap.has(cacheKey)) {
+      return Promise.resolve(cacheMap.get(cacheKey));
+    }
+
+    const resultPromise: Promise<any> = Reflect.apply(target, provider, args);
+    cacheMap.set(cacheKey, resultPromise);
+    return Promise.resolve(resultPromise);
+  };
+
+  return new Proxy(provider, {
+    get: function (target, prop, receiver) {
+      if (prop === 'send') {
+        return function (method, params, isCacheable, subscription) {
+          //caching state_getRuntimeVersion and chain_getHeader
+          //because they are fetched twice per block
+          if (['state_getRuntimeVersion', 'chain_getHeader'].includes(method)) {
+            return cachedMethodHandler(
+              method,
+              params,
+              target.send.bind(target),
+              arguments,
+            );
+          }
+
+          // For other methods, just forward the call to the original method.
+          return Reflect.apply(target.send.bind(target), provider, arguments);
+        };
+      }
+
+      // For other properties, just return the original value.
+      return Reflect.get(target, prop, receiver);
+    },
+  });
+}
